Refactor the common query providers.
Move the `provide` and `provide_extern` functions from
`src/librustc_codegen_llvm/base.rs` to `src/librustc_codegen_util/base.rs`, so
they can be reused in the iron-ox backend. In particular, the
`collect_and_partition_mono_items`, `is_codegened_item`, and `codegen_unit`
queries provided by `provide` are needed by iron-ox
