+++ src/librustc_codegen_ironox/base.rs
@@ -1,6 +1,9 @@
+use rustc::hir::def_id::LOCAL_CRATE;
 use rustc::middle::cstore::EncodedMetadata;
 use rustc::session::config::OutputFilenames;
 use rustc::ty::TyCtxt;
+use rustc::ty::query::Providers;
+use rustc_codegen_utils::base;
 use rustc_data_structures::svh::Svh;
 use syntax_pos::symbol::Symbol;
 
@@ -21,3 +24,12 @@ pub fn codegen_crate<'a, 'tcx>(_tcx: TyCtxt<'a, 'tcx, 'tcx>,
                                -> OngoingCodegen {
     unimplemented!("base::codegen_crate");
 }
+
+pub fn provide(providers: &mut Providers) {
+    base::provide(providers);
+}
+
+pub fn provide_extern(providers: &mut Providers) {
+    base::provide_extern(providers);
+}
+
